packages/justhtml/justhtml-0.15.0.tar.gz/justhtml-0.15.0/tests/justhtml-tests/xml_coercion.dat

#data
#xml-coercion
<!-- normal comment -->
#errors
#document
| <!--  normal comment  -->
| <html>
|   <head>
|   <body>


#data
#xml-coercion
<div>normal text</div>
#errors
#document
| <html>
|   <head>
|   <body>
|     <div>
|       "normal text"


#data
#xml-coercion
<div>端ber</div>
#errors
#document
| <html>
|   <head>
|   <body>
|     <div>
|       "端ber"


#data
#xml-coercion
<div>端ber</div>
#errors
#document
| <html>
|   <head>
|   <body>
|     <div>
|       "端ber "
